requirements.txt
addict==2.4.0
asttokens==3.0.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
cmake==4.0.0
contourpy==1.3.0
cubinlinker-cu11==0.3.0.post2
cuda-bindings==11.8.6
cuda-python==11.8.6
cudf-cu11==24.8.3
# cuml-cu11==24.8.0
cupy-cuda11x==13.4.1
cycler==0.12.1
dask==2024.7.1
dask-cuda==24.8.2
dask-cudf-cu11==24.8.3
dask-expr==1.1.9
decorator==5.2.1
distributed==2024.7.1
distributed-ucxx-cu11==0.39.1
einops==0.8.1
exceptiongroup==1.2.2
executing==2.2.0
fastrlock==0.8.3
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
h5py==3.13.0
huggingface-hub==0.30.2
idna==3.10
importlib-metadata==8.6.1
importlib-resources==6.5.2
ipdb==0.13.13
ipython==8.18.1
jedi==0.19.2
jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
libucx-cu11==1.15.0.post2
lit==18.1.8
llvmlite==0.43.0
locket==1.0.0
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mmcls==0.25.0
mmcv-full==1.5.0
mmsegmentation==0.27.0
mpmath==1.3.0
msgpack==1.1.0
mypy-extensions==1.0.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
nvtx==0.2.11
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.2
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
platformdirs==4.3.7
prettytable==3.16.0
prompt-toolkit==3.0.51
protobuf==6.30.2
psutil==7.0.0
ptxcompiler-cu11==0.8.1.post2
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==16.1.0
pygments==2.19.1
pylibraft-cu11==24.8.1
pynvml==11.4.1
pyparsing==3.2.3
pyre-extensions==0.0.23
python-dateutil==2.9.0.post0
pytz==2025.2
raft-dask-cu11==24.8.1
rapids-dask-dependency==24.8.0
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rmm-cu11==24.8.2
safetensors==0.5.3
scipy==1.13.1
six==1.17.0
sortedcontainers==2.4.0
stack-data==0.6.3
submitit==1.5.2
sympy==1.13.3
tblib==3.1.0
tensorboardx==2.6.2.2
tokenizers==0.21.1
tomli==2.2.1
toolz==1.0.0
torch==2.0.0+cu117
torchmetrics==0.10.3
torchvision==0.15.0+cu117
tornado==6.4.2
traitlets==5.14.3
transformers==4.51.2
treelite==4.3.0
triton==2.0.0
typing-inspect==0.9.0
tzdata==2025.2
ucx-py-cu11==0.39.2
ucxx-cu11==0.39.1
urllib3==2.4.0
wcwidth==0.2.13
xformers==0.0.18
yapf==0.43.0
zict==3.0.0
zipp==3.21.0